+++ apps/infrastructure/src/modules/chatbot/resources.tf
@@ -1,0 +1,4 @@
+resource "random_integer" "ai_kb_bucket_random_integer" {
+  min = 1
+  max = 9999
+}