+++ apps/nextjs-website/src/lib/strapi/fetchFromStrapi.ts
@@ -31,6 +31,7 @@ export const fetchFromStrapi = <A, O, I>(
                 headers: {
                   Authorization: `Bearer ${strapiApiToken}`,
                 },
+                cache: 'no-store',
               }),
             E.toError
           ),