+++ apps/strapi-cms/CHANGELOG.md
@@ -1,5 +1,27 @@
 # strapi-cms
 
+## 1.1.0
+
+### Minor Changes
+
+- ca97805: Update ApiData model
+- 3578c57: Add shortName to Product schema
+
+### Patch Changes
+
+- 9eaa93f: Fix the type of the relations between solutions-products and solutions-webinars
+- bc120e3: Add minimum and maximum values on solution's steps
+
+## 1.0.0
+
+### Major Changes
+
+- 2c05ca2: Refactor homepage schema
+
+### Patch Changes
+
+- 49bc1aa: Add introductionToSteps field to Solution's schema
+
 ## 0.12.0
 
 ### Minor Changes